+++ bin/ocaml_merlin.ml
@@ -29,9 +29,7 @@ let term =
              ouptut.")
   in
   let common = Common.set_print_directory common false in
-  let config =
-    Common.init common ~log_file:No_log_file ~recognize_jbuilder_projects:true
-  in
+  let config = Common.init common ~log_file:No_log_file in
   Scheduler.go ~common ~config (fun () ->
       match dump_config with
       | Some s -> Dune_rules.Merlin_server.dump s
@@ -66,9 +64,7 @@ module Dump_dot_merlin = struct
               "The path to the folder of which the configuration should be \
                printed. Defaults to the current directory.")
     in
-    let config =
-      Common.init common ~log_file:No_log_file ~recognize_jbuilder_projects:true
-    in
+    let config = Common.init common ~log_file:No_log_file in
     Scheduler.go ~common ~config (fun () ->
         match path with
         | Some s -> Dune_rules.Merlin_server.dump_dot_merlin s