+++ test/blackbox-tests/test-cases/syntax-versioning.t/run.t
@@ -9,13 +9,10 @@
   [1]
   $ rm -f dune
 
-  $ echo '(jbuild_version 1)' > jbuild
+Since 3.0.0, jbuild files are plain ignored:
+
+  $ echo 'random stuff' > jbuild
   $ dune build
-  File "jbuild", line 1, characters 0-0:
-  Error: jbuild files are no longer supported, please convert this file to a
-  dune file instead.
-  Note: You can use "dune upgrade" to convert your project to dune.
-  [1]
   $ rm -f jbuild
 
   $ echo '(executable (name x) (link_executables false))' > dune